requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
appdirs==1.4.4
async-timeout==4.0.2
asyncer==0.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
coloredlogs==15.0.1
customtkinter==5.1.3
darkdetect==0.8.0
easygui==0.98.3
fastapi==0.87.0
filetype==1.2.0
flatbuffers==22.12.6
frozenlist==1.3.3
h11==0.14.0
humanfriendly==10.0
idna==3.4
ImageHash==4.3.1
imageio==2.23.0
llvmlite==0.39.1
mpmath==1.2.1
multidict==6.0.3
networkx==2.8.8
numba==0.56.4
numpy==1.23.5
onnxruntime==1.13.1
opencv-python-headless==4.6.0.66
packaging==22.0
pandas==1.5.2
Pillow==9.3.0
pooch==1.6.0
protobuf==4.21.12
pydantic==1.10.2
PyMatting==1.1.8
PyQt6==6.5.1
PyQt6-Qt6==6.5.1
PyQt6-sip==13.5.1
PySide6==6.5.1.1
PySide6-Addons==6.5.1.1
PySide6-Essentials==6.5.1.1
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2022.7
PyWavelets==1.4.1
rembg==2.0.30
requests==2.28.1
scikit-image==0.19.3
scipy==1.9.3
shiboken6==6.5.1.1
six==1.16.0
sniffio==1.3.0
starlette==0.21.0
sympy==1.11.1
Tcl==0.2
tifffile==2022.10.10
tk==0.1.0
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
watchdog==2.1.9
yarl==1.8.2
